requirements.txt
Python 3.8.10
conda==4.10.3
numpy==1.19.5
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1